Release 1.0: Upgrade Handler and Module Improvements (#82)
* Upgrade to 1.0 and fix gosec issues

* Added acltypes.StoreKey to registered upgrade handlers

* Added more store upgrades

* Added remaining custom stores to storeUpgrades

* Removing ccvprovider types store key

* Changed genesis module init order

* use non-consumer modules

* WIP for handling the InitGenesis related error

* manually initialize provider module (#81)

* Renamed upgrade to match go proposal passed in staging

* Renamed upgrade back to 1.0

---------

+++ x/billing/keeper/grpc_query_get_billing_history.go
@@ -36,7 +36,7 @@ func (k Keeper) GetBillingHistory(goCtx context.Context, req *types.QueryGetBill
 		k.cdc.MustUnmarshal(value, &sbhr)
 		// get epoch info
 		epochInfo := k.epochskeeper.GetEpochInfo(ctx, sbhr.EpochIdentifier)
-		epochSince := (epochInfo.CurrentEpoch - int64(sbhr.EpochNumber))
+		epochSince := (epochInfo.CurrentEpoch - sbhr.EpochNumber)
 		epochEventStartTime := epochInfo.CurrentEpochStartTime.Add(-time.Duration(epochSince * int64(epochInfo.Duration)))
 		bhr := types.BillingHistory{
 			ChainletId:        sbhr.ChainletId,
